Aumentando verificação

+++ revisandoBanco/src/main/kotlin/repository/GerenciarCaixas.kt
@@ -6,12 +6,12 @@ import dominio.enums.Material
 import java.math.BigDecimal
 
 val conectar = EntidadeJDBC(
-        url = "jdbc:postgresql://localhost:5433/exemploHuilson",
-        usuario = "postgres",
-        senha = "root"
-    )
+    url = "jdbc:postgresql://localhost:5433/exemploHuilson",
+    usuario = "postgres",
+    senha = "root"
+)
 
-fun criarTabelaCaixa(){
+fun criarTabelaCaixa() {
 
     val sql = "CREATE TABLE IF NOT EXISTS CaixaDAgua (" +
             "id serial NOT NULL PRIMARY KEY, " +
@@ -30,14 +30,14 @@ fun criarTabelaCaixa(){
     banco.close()
 }
 
-fun listarById(){
+fun listarById() {
     var id = ""
     val regex = "^[1-9]\\d*|0\$"
-    do{
+    do {
         println("Digite o ID que deseja listar")
         id = readlnOrNull() ?: "a"
-    }while(id == regex)
-    if(id == "-1"){
+    } while (id == regex)
+    if (id == "-1") {
         println("ID invalido")
         return
     }
@@ -46,7 +46,7 @@ fun listarById(){
     stmt.setInt(1, id.toInt())
     val resp = stmt.executeQuery()
     println("[")
-    while(resp.next()){
+    while (resp.next()) {
         val id = resp.getInt("id")
         val material = resp.getString("material")
         val capacidade = resp.getFloat("capacidade")
@@ -60,9 +60,10 @@ fun listarById(){
 }
 
 fun createCaixa() {
+    println("----------- Criando Caixa -----------")
     /*capacidade: Int,*/
     println("Digite a capacidade da caixa d'água")
-    val capacidade: Int = readln().toIntOrNull() ?:0
+    val capacidade: Int = readln().toIntOrNull() ?: 0
 
     /*material: Material*/
     println("Digite o material da caixa d'água")
@@ -87,11 +88,11 @@ fun createCaixa() {
 
     /*marca: String,*/
     println("Qual a marca da caixa d'água")
-    val marca:String = readlnOrNull() ?: ""
+    val marca: String = readlnOrNull() ?: ""
 
     /*cor: String*/
     println("Qual a cor da caixa d'água")
-    val cor:String = readlnOrNull() ?: ""
+    val cor: String = readlnOrNull() ?: ""
 
     /*tampa: Boolean*/
     println("Tem tampa?(1 = sim, 2 = não)")
@@ -103,7 +104,7 @@ fun createCaixa() {
 
     /*preco: BigDecimal*/
     println("Qual o preço da caixa d'água")
-    val preco:BigDecimal = (readlnOrNull()?.toBigDecimal() ?: "0.0") as BigDecimal
+    val preco: BigDecimal = (readlnOrNull()?.toBigDecimal() ?: "0.0") as BigDecimal
 
     val novaCaixa: CaixaDAgua = CaixaDAgua(
         capacidade = capacidade,
@@ -131,24 +132,42 @@ fun createCaixa() {
     statement.executeUpdate()
     println("Caixa cadastrada com sucesso!")
     banco.close()
+    println("-------------------------------------")
 }
 
 fun editarCaixa() {
+    println("----------- Editando Caixa -----------")
+    val regexOnlyFloats = Regex("^\\d+.\\d+\$")
+    val regexOnlyNumbers = Regex("^\\d+$")
+    val regexEscolhaAtributo = Regex("^[1-7]$")
+    val regexId = Regex("^[1-9][0-9]*$")
     listarCaixas()
-    println("Qual o id da caixa que você deseja editar?")
-    val escolhaId = readln().toIntOrNull() ?: 0
+    var escolhaId = 0
+    do {
+        println("Qual o id da caixa que você deseja editar?")
+        escolhaId = readln().toIntOrNull() ?: 0
+        if (!escolhaId.toString().matches(regexId)) {
+            println("Erro, escolha um id valido")
+        }
+    } while (!escolhaId.toString().matches(regexId))
     val atributo = arrayOf("material", "capacidade", "marca", "cor", "tampa", "preco", "peso")
-    println("1- material, 2-capacidade, 3-marca, 4-cor, 5-tampa, 6-preço, 7-peso")
-    println("Qual atributo deseja trocar?")
+    var atributoEscolhido = "0"
+    do {
+        println("1- material, 2-capacidade, 3-marca, 4-cor, 5-tampa, 6-preço, 7-peso")
+        println("Qual atributo deseja trocar?")
+        atributoEscolhido = readlnOrNull() ?: "0"
+        if (!atributoEscolhido.matches(regexEscolhaAtributo)) {
+            println("Erro, escolha um atributo valido")
+        }
+    } while (!atributoEscolhido.matches(regexEscolhaAtributo))
     var novoAtributo = "";
-    val atributoEscolhido = readlnOrNull() ?: "0"
-    val sql = "UPDATE CaixaDAgua SET ${atributo[atributoEscolhido.toInt()-1]} = ? WHERE id = ?"
+    val sql = "UPDATE CaixaDAgua SET ${atributo[atributoEscolhido.toInt() - 1]} = ? WHERE id = ?"
 
     val banco = conectar.connectarComBanco()
     val stmt = banco!!.prepareStatement(sql)
     stmt.setInt(2, escolhaId)
-    when(atributoEscolhido){
-        "1" ->{
+    when (atributoEscolhido) {
+        "1" -> {
             println("Digite o novo material da caixa d'água")
             println("1 - POLIETILENO")
             println("2 - FIBRADEVIDRO")
@@ -166,24 +185,30 @@ fun editarCaixa() {
             }
             stmt.setString(1, novoAtributo)
         }
-        "2" ->{
-            println("Digite a nova capacidade da caixa d'água")
-            novoAtributo = readlnOrNull() ?: "0"
+
+        "2" -> {
+            do {
+                println("Digite a nova capacidade da caixa d'água")
+                novoAtributo = readlnOrNull() ?: "0"
+            } while (!regexOnlyFloats.matches(novoAtributo))
             stmt.setFloat(1, novoAtributo.toFloat())
         }
-        "3" ->{
+
+        "3" -> {
             println("Qual a nova marca da caixa d'água")
-            novoAtributo  = readlnOrNull() ?: "Sem Marca"
+            novoAtributo = readlnOrNull() ?: "Sem Marca"
 
             stmt.setString(1, novoAtributo)
         }
-        "4" ->{
+
+        "4" -> {
             println("Qual o novo cor da caixa d'água?")
-            novoAtributo = readlnOrNull() ?: "0"
+            novoAtributo = readlnOrNull() ?: "Sem Cor"
 
             stmt.setString(1, novoAtributo)
         }
-        "5" ->{
+
+        "5" -> {
             println("Agora tem tampa?(1 = sim, 2 = não)")
             novoAtributo = when (readln().toInt()) {
                 1 -> "true"
@@ -193,33 +218,42 @@ fun editarCaixa() {
 
             stmt.setBoolean(1, novoAtributo.toBoolean())
         }
-        "6" ->{
-            println("Qual o novo preço da caixa d'água")
-            novoAtributo = readlnOrNull() ?: "0"
+
+        "6" -> {
+            do {
+                println("Qual o novo preço da caixa d'água(Ex: 10.0)")
+                novoAtributo = readlnOrNull() ?: "0.0"
+            } while (!regexOnlyFloats.matches(novoAtributo))
 
             stmt.setString(1, novoAtributo)
         }
-        "7"->{
-            println("Qual o novo peso da caixa?")
-            novoAtributo = readlnOrNull() ?: "0"
 
+        "7" -> {
+            do {
+                println("Qual o novo peso da caixa?")
+                novoAtributo = readlnOrNull() ?: "0"
+            } while (!regexOnlyFloats.matches(novoAtributo))
             stmt.setFloat(1, novoAtributo.toFloat())
         }
-        else ->{
+
+        else -> {
             println("Escolha invalida")
             return
         }
     }
     stmt.executeUpdate()
+    println("Id $escolhaId alterado com sucesso")
     banco.close()
+    println("--------------------------------------")
 }
 
 fun listarCaixas() {
+    println("----------- Listando Caixa(s) -----------")
     val sql = "SELECT * FROM CaixaDAgua"
     val banco = conectar.connectarComBanco()
     val stmt = banco!!.createStatement()
     val resp = stmt.executeQuery(sql)
-    while(resp.next()){
+    while (resp.next()) {
         val id = resp.getInt("id")
         val material = resp.getString("material")
         val capacidade = resp.getFloat("capacidade")
@@ -230,21 +264,39 @@ fun listarCaixas() {
         println("{id = $id, material = $material, capacidade = $capacidade, peso = $peso, cor = $cor, tampa = $tampa, preco = $preco}")
     }
     banco.close()
+    println("-----------------------------------------")
 }
 
 fun deleteCaixas() {
-    println("Qual caixa deseja excluir?(escreva o id)")
+    println("----------- Deletando Caixa -----------")
+    val regexId = Regex("^[1-9][0-9]*$")
+    println("Caixas:[")
     listarCaixas()
-    val caixaExcluir = readln().toInt()
+    println("]")
+    var caixaExcluir = 0
+    do {
+        println("Qual caixa deseja excluir?(escreva o id)")
+        caixaExcluir = readln().toIntOrNull() ?: 0
+        if (!caixaExcluir.toString().matches(regexId)) {
+            println("Erro, escolha um id valido")
+        }
+    } while (!caixaExcluir.toString().matches(regexId))
+    println("Are you sure? Type 1 to execute")
+    if ((readlnOrNull() ?: "0") != "1") {
+        println("Cancelado com sucesso!")
+        return
+    }
+    println("Em processo de eliminação")
     val sql = "DELETE FROM CaixaDAgua WHERE id = ?"
     val banco = conectar.connectarComBanco()
     val stmt = banco!!.prepareStatement(sql)
     stmt.setInt(1, caixaExcluir)
     val linhasAfetadas = stmt.executeUpdate()
-    if(linhasAfetadas > 0){
+    if (linhasAfetadas > 0) {
         println("Excluido com sucesso")
-    }else{
+    } else {
         println("Id não encontrado")
     }
     banco.close()
+    println("---------------------------------------")
 }
